Refactor firebase class to the package Firebase and change name to BatimentDataSource
