+++ app/src/main/java/com/m2dl/mapus/mapus/MainActivity.java
@@ -9,19 +9,16 @@ import android.support.v4.widget.DrawerLayout;
 import android.support.v7.app.ActionBarDrawerToggle;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
-import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
 
 public class MainActivity extends AppCompatActivity
         implements NavigationView.OnNavigationItemSelectedListener {
 
-    private Firebase firebase;
     private Toolbar toolbar;
     
     @Override
     protected void onCreate(Bundle savedInstanceState) {
-        this.firebase = new Firebase();
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
         this.toolbar = (Toolbar) findViewById(R.id.toolbar);
